+++ src/main/java/hotel/entity/Admins.java
@@ -23,6 +23,6 @@ public class Admins {
     private String password;
     private String card_number;
     private String valid_date;
-    private Float balance;
+    private Double balance;
     private Date created_at;
 }
